+++ packages/renderer/src/lib/chat/components/chat-header.svelte
@@ -6,7 +6,7 @@ import { router } from 'tinro';
 
 import type { ModelInfo } from '/@/lib/chat/components/model-info';
 import ModelSelector from '/@/lib/chat/components/model-selector.svelte';
-import { currentChatId } from '/@/lib/chat/stores/current-chat-id.svelte.js';
+import { currentChatId } from '/@/lib/chat/state/current-chat-id.svelte.js';
 import { mcpRemoteServerInfos } from '/@/stores/mcp-remote-servers';
 import type { MCPRemoteServerInfo } from '/@api/mcp/mcp-server-info';
