+++ packages/renderer/src/lib/chat/components/sidebar-history/item.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 import { router } from 'tinro';
 
-import { currentChatId } from '/@/lib/chat/stores/current-chat-id.svelte.js';
-
 import type { Chat } from '../../../../../../main/src/chat/db/schema';
+import { currentChatId } from '../../state/current-chat-id.svelte.js';
 import MoreHorizontalIcon from '../icons/more-horizontal.svelte';
 import TrashIcon from '../icons/trash.svelte';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '../ui/dropdown-menu';
