+++ packages/renderer/src/lib/chat/state/sidebar-collapsed.svelte.ts
